medical_agent/requirements.txt
langchain==0.3.23
langchain_core==0.3.53
langchain_groq==0.3.2
mistralai==1.7.0
python-dotenv==1.1.0
